refactor(v2): separate token budget policy
