+++ instructor/v2/core/retry.py
@@ -6,10 +6,8 @@ instead of v1's process_response.
 
 from __future__ import annotations
 
-import copy
 import json
 import logging
-from numbers import Real
 from typing import TYPE_CHECKING, Any, TypeVar
 
 from pydantic import BaseModel, ValidationError
@@ -23,6 +21,10 @@ from tenacity import (
 
 from instructor.v2.core.mode import Mode
 from instructor.v2.core.providers import Provider
+from instructor.v2.core.budget import (
+    _budget_error as _budget_error,
+    _validate_token_budget as _validate_token_budget,
+)
 from instructor.v2.core.errors import (
     AsyncValidationError,
     FailedAttempt,
@@ -30,14 +32,17 @@ from instructor.v2.core.errors import (
     InstructorRetryException,
     ResponseParsingError,
     TokenBudgetError,
-    TokenBudgetExceeded,
-    TokenUsageUnavailableError,
 )
 from instructor.v2.dsl.iterable import IterableBase
 from instructor.v2.dsl.response_list import ListResponse
 from instructor.v2.dsl.simple_type import AdapterBase
 from instructor.v2.core.messages import extract_messages
-from instructor.v2.core.usage import has_compatible_usage, update_total_usage
+from instructor.v2.core.usage import (
+    _usage_snapshot as _usage_snapshot,
+    _usage_total_tokens as _usage_total_tokens,
+    has_compatible_usage,
+    update_total_usage,
+)
 from instructor.v2.core.exceptions import RegistryValidationMixin
 from instructor.v2.core.registry import mode_registry
 
@@ -74,50 +79,6 @@ def _attempt_metadata(
     }
 
 
-def _usage_snapshot(total_usage: Any) -> Any:
-    if isinstance(total_usage, BaseModel):
-        return total_usage.model_copy(deep=True)
-    return copy.deepcopy(total_usage)
-
-
-def _usage_total_tokens(total_usage: Any) -> int | None:
-    direct_total = getattr(total_usage, "total_tokens", None)
-    if isinstance(direct_total, Real) and not isinstance(direct_total, bool):
-        return int(direct_total)
-
-    token_fields = (
-        "input_tokens",
-        "output_tokens",
-        "cache_creation_input_tokens",
-        "cache_read_input_tokens",
-    )
-    values = [getattr(total_usage, field, None) for field in token_fields]
-    numeric_values = [
-        int(value)
-        for value in values
-        if isinstance(value, Real) and not isinstance(value, bool)
-    ]
-    return sum(numeric_values) if numeric_values else None
-
-
-def _validate_token_budget(
-    token_budget: int | None,
-    *,
-    response_model: object,
-    kwargs: dict[str, Any],
-) -> None:
-    if token_budget is None:
-        return
-    if isinstance(token_budget, bool) or not isinstance(token_budget, int):
-        raise TypeError("token_budget must be a positive integer or None")
-    if token_budget <= 0:
-        raise ValueError("token_budget must be greater than zero")
-    if response_model is None:
-        raise ValueError("token_budget requires a structured response_model")
-    if kwargs.get("stream"):
-        raise ValueError("token_budget is not supported for streaming responses")
-
-
 def _finalize_parsed_response(
     parsed: Any,
     response: Any,
@@ -143,50 +104,6 @@ def _finalize_parsed_response(
         if usage is not None:
             parsed._total_usage = usage  # type: ignore[attr-defined]
     return parsed
-
-
-def _budget_error(
-    *,
-    token_budget: int | None,
-    usage_available: bool,
-    total_usage: Any,
-    attempt_number: int,
-    response: Any,
-    kwargs: dict[str, Any],
-    failed_attempts: list[FailedAttempt],
-) -> TokenBudgetError | None:
-    if token_budget is None:
-        return None
-
-    error_kwargs = {
-        "budget": token_budget,
-        "last_completion": response,
-        "messages": extract_messages(kwargs),
-        "n_attempts": attempt_number,
-        "total_usage": _usage_snapshot(total_usage),
-        "create_kwargs": kwargs,
-        "failed_attempts": failed_attempts,
-    }
-    if not usage_available:
-        return TokenUsageUnavailableError(
-            "Token budget cannot be enforced because the provider response did "
-            "not include compatible usage metadata",
-            **error_kwargs,
-        )
-
-    used_tokens = _usage_total_tokens(total_usage)
-    if used_tokens is None:
-        return TokenUsageUnavailableError(
-            "Token budget cannot be enforced because total token usage is unavailable",
-            **error_kwargs,
-        )
-    if used_tokens >= token_budget:
-        return TokenBudgetExceeded(
-            f"Token budget exhausted after {used_tokens} tokens across "
-            f"{attempt_number} attempts (budget: {token_budget})",
-            **error_kwargs,
-        )
-    return None
 
 
 def _initialize_usage(provider: Provider | Mode) -> Any:
